move 'pullDataFromSftp' into it's own dir 'ingestCSV'
